Extract notebook content so it can be used in other places
