+++ src/components/notebook/NotebookViewer.tsx
@@ -34,6 +34,7 @@ const LazyDebugPanel = React.lazy(() =>
 import { CellAdder } from "./cell/CellAdder.js";
 import { EmptyStateCellAdder } from "./EmptyStateCellAdder.js";
 import { GitCommitHash } from "./GitCommitHash.js";
+import { NotebookContent } from "./NotebookContent.js";
 
 interface NotebookViewerProps {
   notebookId: string;
@@ -61,24 +62,7 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
 
   const [showRuntimeHelper, setShowRuntimeHelper] = React.useState(false);
 
-  const focusedCellId = useQuery(focusedCellSignal$);
   const contextSelectionMode = useQuery(contextSelectionMode$);
-  const hasManuallyFocused = useQuery(hasManuallyFocused$);
-
-  // Reset focus when focused cell changes or is removed
-  React.useEffect(() => {
-    if (focusedCellId && !cellReferences.find((c) => c.id === focusedCellId)) {
-      store.setSignal(focusedCellSignal$, null);
-    }
-  }, [focusedCellId, cellReferences, store]);
-
-  // Focus first cell when notebook loads and has cells (but not after deletion)
-  React.useEffect(() => {
-    if (!focusedCellId && cellReferences.length > 0 && !hasManuallyFocused) {
-      store.setSignal(focusedCellSignal$, cellReferences[0].id);
-      store.setSignal(hasManuallyFocused$, true);
-    }
-  }, [focusedCellId, cellReferences, store, hasManuallyFocused]);
 
   // cells are already sorted by position from the database query
 
@@ -286,24 +270,7 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
             )}
 
             {/* Cells */}
-            {cellReferences.length === 0 ? (
-              <EmptyStateCellAdder />
-            ) : (
-              <>
-                <ErrorBoundary fallback={<div>Error rendering cell list</div>}>
-                  <CellList cellReferences={cellReferences} />
-                </ErrorBoundary>
-                {/* Add Cell Buttons */}
-                <div className="border-border/30 mt-6 border-t px-4 pt-4 sm:mt-8 sm:px-0 sm:pt-6">
-                  <div className="space-y-3 text-center">
-                    <CellAdder position="after" />
-                    <div className="text-muted-foreground mt-2 hidden text-xs sm:block">
-                      Add a new cell
-                    </div>
-                  </div>
-                </div>
-              </>
-            )}
+            <NotebookContent />
           </div>
         </div>
