+++ src/main.tsx
@@ -5,11 +5,14 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
 import { App } from "./Root.js";
+import { DebugProvider } from "./debug-mode.js";
 
 ReactDOM.createRoot(document.getElementById("react-app")!).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <DebugProvider>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </DebugProvider>
   </React.StrictMode>
 );
